Introduce module and "command::definition" class
